EventBroadcaster thread-safe (phase 10): Major refactoring to enable reuse of ConcurrentResourceFloodTest objects
